ALSA: timer: remove usage of list iterator for list_add() after the list body
In preparation to limit the scope of a list iterator to the list
traversal loop, use a dedicated pointer to point to the found element.

Before, the code implicitly used the head when no element was found
when using &pos->list. Since the new variable is only set if an
element was found, the list_add() is explicitly performed on the list
variable when set and on the list head otherwise.

Signed-off-by: Jakob Koschel <jakobkoschel@gmail.com>