powerpc: spufs: remove unnecessary use of list iterator after loop
The list iterator variable is still set even if the list is empty and
holds a bogus ptr computed based on the head of the list.

In preparation to limiting the scope of a list iterator to the list
traversal loop we avoid using it after the loop altogether by moving
the relevant code into the loop.

This makes sure the code only executes on valid list iterator variables.

This is part of the effort to limit the list iterator variable to the
loop body [1].

Link: https://lore.kernel.org/all/CAHk-=wgRr_D8CB-D9Kg-c=EHreAsk5SqXPwr9Y7k9sA6cWXJ6w@mail.gmail.com/
Signed-off-by: Jakob Koschel <jakobkoschel@gmail.com>