team: Remove usage of list iterator variable for
In preparation to limit the scope of the list iterator variable to the
list traversal loop, use a dedicated pointer to iterate through the list.

Since that variable should not be used past the loop iteration, a
separate variable is used to 'remember the current location within the
loop'.

To either continue iterating from that position or skip the iteration
(if the previous iteration was complete) list_prepare_entry() is used.

Signed-off-by: Jakob Koschel <jakobkoschel@gmail.com>